Pretty much finished the assignment im so tired i can't continue
Added the health function for the bones since theres 3 stages the user has to hit for them to vanish.

Made it so that if you touch the bot bounds player loses a life.

This mini assignment is filling me with hatred i think i will have to leave it half finished. I created a framework that i can build off of in the future though.